requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-magic-bin==0.4.14; sys_platform == "win32"
python-magic==0.4.27; sys_platform != "win32"
clamd==1.0.2